requirements.txt
streamlit==1.31.0
langchain==0.1.8
langchain-google-genai==0.0.7
neo4j==5.15.0
python-dotenv==1.0.0
google-generativeai==0.3.1
